requirements.txt
python-telegram-bot==20.7
httpx==0.24.1
sentence-transformers
sqlite-utils
numpy
wikipedia